690 with 681 (#909)
Passed sanity
* BEO-10702 Unable to create/update subscription

* BEO-10350 Implement watch history threshold for linear channels (#841)

* Update Web.config

* BEO-10606 Protect basic fields in Ingest V2 (#861)

* 680 beo 10472 update is active on cancelation (#850)

* BEO-10472 update is active on cancelation

* add is active to subscription

* fix - BEO-10686

* reflector for ImageTypeName

* Retrieve all currencies no matter what! (#877)

* 681 beo 10272 manual collections (#840)

* BEO-10272 add AssetTYpe column

* BEO-10272 add/update/get

* no message

* Adding permissions

* BEO-10272 - asset/action/list search bu cannel

* BEO-10272 KalturaManualCollectionAssetType & KalturaManualCollectionAsset

* fix BEO-10697


* 680 beo 10741 null dt row check (#883)

* WIP BEO-10741

* Fix permission validation in  ottUser-list - validate partner permissions instead of hard coded operator (#884)

* fix BEO-10725 "cb_max_insert_tries", 2

* fix channelsIdsWithMetadata

* BEO-10532 Add missed fields from EpgAsset and EpgCB to each other during mapping. (#874)

* remove old device play data after re adding device to domain (#885)

* remove old device play data after re adding device to domain

* remove old device play data after re adding device to domain

* fix update of KalturaManualCollectionAsset

* 681 add logs for GEO block in AssetRules (#890)

* 680 beo 10485 inheritable roles (#848)

* 680_BEO-10485_Inheritable_roles


* BEO-10296 Region improvements (partial updates). (#793)

* BEO-10296 Initial commit.

* BEO-10296 Code review refactoring.

* BEO-10296 Add feature flag (TCM) to turn on/off improvement for region update.

* BEO-10296 Fix according to rebase.

* 681 fix update manual coll (#899)

* fix updateManualColl

* fix updateManualColl

* fix BEO-10794 (#903)

+++ RemoteTasks/ElasticsearchHandlerCommon/Utils.cs
@@ -152,7 +152,7 @@ namespace ElasticsearchTasksCommon
             return sRules;
         }
 
-        public static EpgCB GetEpgProgram(int nGroupID, int nEpgID)
+        public static EpgCB GetEpgProgram(int nGroupID, long nEpgID)
         {
             EpgCB res = null;
 
@@ -232,7 +232,7 @@ namespace ElasticsearchTasksCommon
             return res;
         }
 
-        public static List<EpgCB> GetEpgPrograms(int groupId, int epgId, List<string> languages, EpgBL.BaseEpgBL epgBL = null)
+        public static List<EpgCB> GetEpgPrograms(int groupId, long epgId, List<string> languages, EpgBL.BaseEpgBL epgBL = null)
         {
             List<EpgCB> results = new List<EpgCB>();
